AUT-426: Update to use dynamic version instead of hardcoded value
This change updates Gradle to use BUILD_NUMBER as an environment variable instead of the hardcoded value "SNAPSHOT". Whenever there is a new commit added to the master branch, the build number increments by 1 provided that the build is successful.

The version will start from v1.3.1 instead of v1.2-SNAPSHOT.

Author: @adityapahuja